updated ipc to work in more generic conexts, use generic messages, and send is now Send using FutureExt::race
